pyproject.toml
[project]
name = "mangahub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "bs4>=0.0.2",
    "easy-config-hub",
    "httpx>=0.28.1",
    "icecream>=2.1.4",
    "loguru>=0.7.3",
    "lxml>=5.3.1",
    "nuitka>=2.6.7",
    "pillow>=11.0.0",
    "pip>=25.0.1",
    "pyan3>=1.2.0",
    "pydantic>=2.10.4",
    "pyinstaller>=6.11.1",
    "pyside6>=6.8.1",
    "pysnooper>=1.2.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tomli>=2.2.1",
]

[tool.mypy]
python = "3.13"
ignore_missing_imports = true
strict = false

[tool.uv.sources]
easy-config-hub = { path = "../packages/easy-config-hub" }
